requirements.txt
opencv-python
numpy
matplotlib
ipykernel
facenet_pytorch
torch
torchvision
pandas
pillow
openpyxl
streamlit
dlib-19.24.99-cp313-cp313-win_amd64.whl